Update
in the last if statement
"If they do not slap the card:
         the next player will take their turn"
the if statement shouldn't be there because regardless of whether or not they slap the card, the next player will take their turn. I've made the changes in all other similar instances, so remember that nothing is implied in coding.